migrationファイルのplansテーブルの修正３

+++ ikkyu/database/migrations/2022_05_19_022756_create_plans_table.php
@@ -16,7 +16,7 @@ class CreatePlansTable extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table->string('name', 100);
-            $table->integer('hotel_id')->unsigned()->index();
+            $table->bigInteger('hotel_id')->unsigned()->index();
             $table->integer('price')->unsigned();
             $table->integer('rooms')->unsigned();;
             $table->timestamps();
